Update R237B_Ive-w31p9hEm2-3.krn
I worked out the tremolo issue for bass drum and violins in bar 1: I deleted the ties.
It looks better now, but I was unable to add a tie for the timpani bars 1 and 2.
It was not possible for bar 2. There is a hanging tie I could not find to delete...
I was unable to fix the tremolo of cello div 2 bar 2.