Update R218_Sch-w33p200m252-256.krn
Grouped Soprano with strings (grp:A) based on rhythmic pattern. Grouped m. 140 for mandolin, piano and harp with soprano and strings (grp:A) based on higher number of matching pitches. This example is difficult because the voice parts are overlapping and the mandolin, piano, and harp have no phrasing, and double each voice part in turn.